+++ src/imio/smartweb/core/sharing/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+import imio.smartweb.core.sharing.permissions